+++ backend/apps/owasp/graphql/nodes/chapter.py
@@ -2,8 +2,7 @@
 
 import graphene
 
-from apps.common.graphql.nodes import BaseNode
-from apps.github.graphql.nodes.user import UserNode
+from apps.owasp.graphql.nodes.common import GenericEntityNode
 from apps.owasp.models.chapter import Chapter
 
 
@@ -14,17 +13,12 @@ class GeoLocationType(graphene.ObjectType):
     lng = graphene.Float()
 
 
-class ChapterNode(BaseNode):
+class ChapterNode(GenericEntityNode):
     """Chapter node."""
 
     key = graphene.String()
-    related_urls = graphene.List(graphene.String)
     suggested_location = graphene.String()
-    updated_at = graphene.Float()
-    url = graphene.String()
-
     geo_location = graphene.Field(GeoLocationType)
-    top_contributors = graphene.List(UserNode)
 
     class Meta:
         model = Chapter
@@ -47,22 +41,6 @@ class ChapterNode(BaseNode):
         """Resolve chapter key."""
         return self.idx_key
 
-    def resolve_related_urls(self, info):
-        """Resolve chapter related urls."""
-        return self.idx_related_urls
-
     def resolve_suggested_location(self, info):
         """Resolve chapter suggested location."""
         return self.idx_suggested_location
-
-    def resolve_top_contributors(self, info):
-        """Resolve chapter top contributors."""
-        return [UserNode(**repo) for repo in self.idx_top_contributors]
-
-    def resolve_updated_at(self, info):
-        """Resolve updated at."""
-        return self.idx_updated_at
-
-    def resolve_url(self, info):
-        """Resolve url."""
-        return self.idx_url
